% summary.pedigree.Rd
%--------------------------------------------------------------------------
% What: Extend pedigree man page
% $Id: summary.pedigree.Rd 1092 2006-11-12 12:28:38Z ggorjan $
% Time-stamp: <2006-09-30 01:32:11 ggorjan>
%--------------------------------------------------------------------------

\name{summary.Pedigree}

\alias{summary.Pedigree}

\title{Pedigree summary}

\description{\code{summary.Pedigree} reports TODO.
}

\usage{\method{summary}{Pedigree}(object, \ldots)}

\arguments{
  \item{object}{pedigree object}
  \item{\ldots}{additional arguments for other methods (not used)}
}

\details{
  TODO.
}

\value{
  TODO.
}

\author{Gregor Gorjanc}

\seealso{\code{\link{Pedigree}}}

\examples{
  ped <- generatePedigree(nId=5)
  summary(ped)
}

\keyword{manip}
\keyword{misc}

%--------------------------------------------------------------------------
% summary.pedigree.Rd ends here
