% undocumented.Rd
%--------------------------------------------------------------------------
% What: Undocumented functions
% $Id: undocumented.Rd 1092 2006-11-12 12:28:38Z ggorjan $
% Time-stamp: <2006-09-30 01:31:29 ggorjan>
%--------------------------------------------------------------------------

\name{undocumented}

\alias{undocumented}

\title{Undocumented functions}

\description{

These functions are undocumented. Some are internal and not intended for
direct use. Some are not yet ready for end users. Others simply haven't
been documented yet.

}

\keyword{misc}

%--------------------------------------------------------------------------
% undocumented.Rd ends here
